+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Todas as mudanças de destaque nesse projeto será documentado neste arquivo. 
 
+## [Não publicado]
+
+### Modificado
+- Nova estrutura de JSON para o cache, que mantém a ordem dos prêmios retornado pelo parser.
+
 ## [0.0.5] - 2017-12-04
 
 ### Adicionado